fix(agent): MCP tool sweep — 4 integration bugs
Surfaced by a synthetic call_tool() sweep against evolat:

- list_accounts / get_account: removed references to a non-existent
  account_type field — Account uses report_category for taxonomy.
  Every list_accounts call was raising FieldError.
- suggest_reconciliation: 1.0 - (Decimal / Decimal) raised TypeError
  in _find_existing_book_matches confidence scoring. Cast the ratio
  to float before subtracting from 1.0. This was blocking the agent's
  reconciliation-suggestion path entirely.
- financial_statements: the agent-facing trim keyed off "name" / "total"
  but the underlying pipeline emits "label" / "amount". Result: every
  category surfaced as {name: null, total: null}. Switched to correct
  keys and added period / basis / account_count to the response.
- financial_statements: bad date strings (and reversed ranges) now
  return a clean {"error": ...} blob instead of bubbling ValueError
  out to the agent loop.

+++ accounting/services/bank_transaction_suggestion_service.py
@@ -402,8 +402,12 @@ class BankTransactionSuggestionService:
             if abs(difference) > bank_amount * 2:  # Don't suggest if difference is more than 2x the bank amount
                 continue
             
-            # Calculate confidence based on similarity and amount match
-            amount_match_score = 1.0 - (abs(difference) / bank_amount) if bank_amount > 0 else 0.0
+            # Calculate confidence based on similarity and amount match.
+            # Cast to float — `difference` and `bank_amount` are Decimals;
+            # mixing 1.0 (float) with Decimal raises TypeError.
+            amount_match_score = (
+                1.0 - float(abs(difference) / bank_amount) if bank_amount > 0 else 0.0
+            )
             amount_match_score = max(0.0, min(1.0, amount_match_score))
             
             confidence = (0.6 * similarity) + (0.4 * amount_match_score)
